refactor(muya): extract ImageSelection and cursorCoords from Selection
